Altered the cmake scripts to prevent absolute paths from ending up in the generated configuration file.  Absolute paths prevent the package from being "installable".
This fix requires that the build interface and install interface for the include directories be cconfigured such that the install path is relative.  This requires moving the handling of some include path element over to a modern cmake way of handling them, i.e. via targets.

To remedy the lib dependencies being absolute paths requires using the OpenEXR target rather than the library file names.